rename package util to internal/assert
