+++ internal/assert/assert.go
@@ -1,4 +1,4 @@
-package util
+package assert
 
 import (
 	"fmt"
@@ -10,9 +10,9 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-// AssertDeepCopyEqual checks to see if two variables have the same values but DO NOT share any memory
+// DeepCopyEqual checks to see if two variables have the same values but DO NOT share any memory
 // There is currently a special case for `time.loc` (as this code traverses into unexported fields)
-func AssertDeepCopyEqual(t *testing.T, a interface{}, b interface{}) {
+func DeepCopyEqual(t *testing.T, a interface{}, b interface{}) {
 	v1 := reflect.ValueOf(a)
 	v2 := reflect.ValueOf(b)
